Implement Singleton class and refactor view controller code
- Implement Singleton class(Goals, Wishes, Givingups)
- Add table count constraints about 5 goals, 25 total number
- Refactor code

+++ FiveTwoFiveRule/Controller/WishViewController.swift
@@ -14,59 +14,27 @@ class WishViewController: UIViewController {
     @IBOutlet weak var wishTableView: UITableView!
     
     let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
-    var wishes:[Wish]?
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        NotificationCenter.default.addObserver(self, selector: #selector(fetchWish), name: Notification.Name("FetchWish"), object: nil)
-
+        NotificationCenter.default.addObserver(self, selector: #selector(reload), name: Notification.Name("ReloadWish"), object: nil)
+        
         fetchWish()
     }
-
-    @IBAction func touchUpAddWishButton(_ sender: UIBarButtonItem) {
-        let alert = UIAlertController(title: "Add Wish", message: "Enter your wish you think seriously.", preferredStyle: .alert)
-        alert.addTextField()
-        
-        /// To do: Exception handling for null string input
-        let submitButton = UIAlertAction(title: "Add", style: .default, handler: { (action) in
-            let textField = alert.textFields![0]
-            
-            let newWish = Wish(context: self.context)
-            newWish.name = textField.text
-            
-            do {
-                try self.context.save()
-            }
-            catch {
-                
-            }
-            
-            /// To do: Change fetchWish to inserttableview for smooth adding animation
-            self.fetchWish()
-            
-        })
-        
-        let cancelButton = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-        
-        alert.addAction(submitButton)
-        alert.addAction(cancelButton)
-        
-        self.present(alert, animated: true, completion: nil)
-    }
-
+    
 }
 
 extension WishViewController: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return self.wishes?.count ?? 0
+        return Wishes.shared.wishes.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = wishTableView.dequeueReusableCell(withIdentifier: "WishCell", for: indexPath)
         
-        let wish = self.wishes![indexPath.row]
+        let wish = Wishes.shared.wishes[indexPath.row]
         
         cell.textLabel?.text = wish.name
         
@@ -79,24 +47,31 @@ extension WishViewController: UITableViewDelegate {
     
     func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let goalSwipeAction = UIContextualAction(style: .destructive, title: "Goal") { (action, view, completion) in
+            guard Goals.shared.goals.count < 5 else {
+                // To do: Show Alert
+                return
+            }
+            
             // Move wish to goal
-            let wishToGoal = self.wishes![indexPath.row]
+            let wishToGoal = Wishes.shared.wishes[indexPath.row]
+            
             let goalFromWish = Goal(context: self.context)
             goalFromWish.name = wishToGoal.name
+            
             self.context.delete(wishToGoal)
+            
             do {
                 try self.context.save()
+            
+                Goals.shared.goals.append(goalFromWish)
+                NotificationCenter.default.post(name: Notification.Name("ReloadGoal"), object: nil)
+                
+                Wishes.shared.wishes.remove(at: indexPath.row)
+                self.wishTableView.deleteRows(at: [indexPath], with: UITableView.RowAnimation.automatic)
             }
             catch {
                 /// To do:
             }
-            
-            // Post Notification for fetching goal table view
-            NotificationCenter.default.post(name: Notification.Name("FetchGoal"), object: nil)
-            
-            // Delete wish in wish table view
-            self.wishes?.remove(at: indexPath.row)
-            self.wishTableView.deleteRows(at: [indexPath], with: UITableView.RowAnimation.automatic)
         }
         
         goalSwipeAction.backgroundColor = UIColor.systemGreen
@@ -107,23 +82,25 @@ extension WishViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, leadingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         let givingupSwipeAction = UIContextualAction(style: .destructive, title: "Giving-up") { (action, view, completion) in
             // Move wish to giving-up
-            let wishToGivingup = self.wishes![indexPath.row]
+            let wishToGivingup = Wishes.shared.wishes[indexPath.row]
+            
             let givingupFromWish = Givingup(context: self.context)
             givingupFromWish.name = wishToGivingup.name
+            
             self.context.delete(wishToGivingup)
+            
             do {
                 try self.context.save()
+                
+                Givingups.shared.givingups.append(givingupFromWish)
+                NotificationCenter.default.post(name: Notification.Name("ReloadGivingup"), object: nil)
+                
+                Wishes.shared.wishes.remove(at: indexPath.row)
+                self.wishTableView.deleteRows(at: [indexPath], with: UITableView.RowAnimation.automatic)
             }
             catch {
                 /// To do:
             }
-            
-            // Post Notification for fetching givingup table view
-            NotificationCenter.default.post(name: Notification.Name("FetchGivingup"), object: nil)
-            
-            // Delete wish in wish table view
-            self.wishes?.remove(at: indexPath.row)
-            self.wishTableView.deleteRows(at: [indexPath], with: UITableView.RowAnimation.automatic)
         }
         
         givingupSwipeAction.backgroundColor = UIColor.systemRed
@@ -135,9 +112,22 @@ extension WishViewController: UITableViewDelegate {
 
 extension WishViewController {
     
-    @objc func fetchWish() {
+    @IBAction func touchUpAddWishButton(_ sender: UIBarButtonItem) {
+        guard (Goals.shared.goals.count + Wishes.shared.wishes.count + Givingups.shared.givingups.count) < 25 else {
+            // To do: Show Alert
+            return
+        }
+        
+        presentAddWishAlert()
+    }
+    
+}
+
+extension WishViewController {
+    
+    func fetchWish() {
         do {
-            self.wishes = try self.context.fetch(Wish.fetchRequest())
+            Wishes.shared.wishes = try self.context.fetch(Wish.fetchRequest())
             
             DispatchQueue.main.async {
                 self.wishTableView.reloadData()
@@ -146,6 +136,46 @@ extension WishViewController {
         catch {
             
         }
+    }
+    
+    @objc func reload() {
+        DispatchQueue.main.async {
+            self.wishTableView.reloadData()
+        }
+    }
+    
+    func presentAddWishAlert() {
+         let alert = UIAlertController(title: "Add Wish", message: "Enter your wish you think seriously.", preferredStyle: .alert)
+         alert.addTextField()
+         
+         /// To do: Exception handling for null string input
+         let submitButton = UIAlertAction(title: "Add", style: .default, handler: { (action) in
+             let textField = alert.textFields![0]
+             
+             let wish = Wish(context: self.context)
+             wish.name = textField.text
+             
+             do {
+                 try self.context.save()
+                 
+                 Wishes.shared.wishes.append(wish)
+                 
+                 // Insert Wish Cell
+                 self.wishTableView.beginUpdates()
+                 self.wishTableView.insertRows(at: [IndexPath(row: Wishes.shared.wishes.count - 1, section: 0)], with: UITableView.RowAnimation.none)
+                 self.wishTableView.endUpdates()
+             }
+             catch {
+                 
+             }
+         })
+         
+         let cancelButton = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
+         
+         alert.addAction(submitButton)
+         alert.addAction(cancelButton)
+         
+         self.present(alert, animated: true, completion: nil)
     }
     
 }
